Python/Django/test_api/my_project/mysite/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
djangorestframework==3.6.3
idna==2.6
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
urllib3==1.22
